Inline `handle_model` into `process_model`
